+++ src/pages/signup/_components/MailAuthActivity.tsx
@@ -10,7 +10,7 @@ import {
 } from "@/components/ui/form";
 import { Button } from "@/components/ui/button";
 
-import { useStackForm } from "../_hooks/useStackForm";
+import { useStackForm } from "../../../hooks/useStackForm";
 import NextStepButton from "./NextStepButton";
 import {
   Activity,
