+++ src/pages/signup/_components/NextStepButton.tsx
@@ -2,7 +2,7 @@ import { Button } from "@/components/ui/button";
 
 import { useMyFlow } from "@/utils/useMyFlow";
 
-import { FormType } from "../_hooks/useStackForm";
+import { FormType } from "../../../hooks/useStackForm";
 
 import { useNavigate } from "@/router";
