+++ src/pages/signup/_components/UserTypeActivity.tsx
@@ -11,8 +11,8 @@ import {
 } from "@/components/ui/form";
 import Logo from "@/components/Logo";
 
-import { validate } from "../_utils/vaildate";
-import { useStackForm } from "../_hooks/useStackForm";
+import { validateForm } from "../../../utils/vaildateForm";
+import { useStackForm } from "../../../hooks/useStackForm";
 import UserTypeItem from "./UserTypeItem";
 import NextStepButton from "./NextStepButton";
 import {
@@ -92,7 +92,7 @@ const UserTypeActivity: ActivityComponentType = () => {
                       activityName={"NameActivity" as never}
                       params={{ form }}
                       disabled={
-                        !validate({
+                        !validateForm({
                           type: "type",
                           value: field.value,
                         })
